+++ packages/react-instantsearch/src/widgets/Chat.tsx
@@ -310,6 +310,7 @@ function ChatInner<
       messagesProps={{
         status,
         onReload: (messageId) => regenerate({ messageId }),
+        onNewConversation: clearMessages,
         onClose: () => setOpen(false),
         sendMessage: sendMessage as ChatUiProps['sendMessage'],
         setInput,
@@ -342,6 +343,7 @@ function ChatInner<
         translations: messagesTranslations,
         messageTranslations,
         ...messagesProps,
+        error,
       }}
       promptProps={{
         promptRef,