+++ packages/instantsearch.js/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.101.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.100.0...instantsearch.js@4.101.0) (2026-06-03)
+
+
+### Features
+
+* **chat:** generic error handling ([#6997](https://github.com/algolia/instantsearch/issues/6997)) ([c307370](https://github.com/algolia/instantsearch/commit/c307370479f5764430160e4d3d007463ecd5e361))
+
+
+
+
+
+# [4.100.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.99.0...instantsearch.js@4.100.0) (2026-06-02)
+
+
+### Features
+
+* **chat:** split chat trigger widget ([#6966](https://github.com/algolia/instantsearch/issues/6966)) ([379084c](https://github.com/algolia/instantsearch/commit/379084cd459a0d8b0f4f2a55b0dae336cf331008))
+
+
+
+
+
 # [4.99.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.98.0...instantsearch.js@4.99.0) (2026-05-26)