+++ packages/react-instantsearch-nextjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.7](https://github.com/algolia/instantsearch/compare/react-instantsearch-nextjs@1.2.6...react-instantsearch-nextjs@1.2.7) (2026-06-03)
+
+**Note:** Version bump only for package react-instantsearch-nextjs
+
+
+
+
+
+## [1.2.6](https://github.com/algolia/instantsearch/compare/react-instantsearch-nextjs@1.2.5...react-instantsearch-nextjs@1.2.6) (2026-06-02)
+
+**Note:** Version bump only for package react-instantsearch-nextjs
+
+
+
+
+
 ## [1.2.5](https://github.com/algolia/instantsearch/compare/react-instantsearch-nextjs@1.2.4...react-instantsearch-nextjs@1.2.5) (2026-05-26)
 
 **Note:** Version bump only for package react-instantsearch-nextjs